+++ src/main/java/com/kevinbooms/App.java
@@ -8,13 +8,20 @@ public class App {
 
         System.out.println("Let's build Taskmin and finally finish it!\n");
 
-        // manually testing Task and ToDoList class
-        Task taskOne = new Task(1, "The first task");
-        Task taskTwo = new Task(2, "The next right thing to do.");
-        ToDoList list = new ToDoList();
-        list.addTask(taskOne);
-        list.addTask(taskTwo);
+        CommandLine cl = new CommandLine();
+        System.out.println(cl.printGreeting());
+        cl.displayBanner();
+        cl.displayMainMenu();
+        System.out.println("*** List menu sample ***");
+        cl.displayListMenu();
 
-        System.out.println(list.getTasks().get(0).getDescription());
+        // *** manually testing Task and ToDoList class ***
+//        Task taskOne = new Task(1, "The first task");
+//        Task taskTwo = new Task(2, "The next right thing to do.");
+//        ToDoList list = new ToDoList();
+//        list.addTask(taskOne);
+//        list.addTask(taskTwo);
+//
+//        System.out.println(list.getTasks().get(0).getDescription());
     }
 }